[SELS-TASK][UI/BE] User Category List V2

+++ back-end/e-learning/app/Http/Controllers/WordController.php
@@ -94,8 +94,7 @@ class WordController extends Controller
 				}
 			}
 			else {
-				if (count($choices) >= $x)
-					$choices[$x]->forceDelete();
+				if (count($choices) > $x) $choices[$x]->forceDelete();
 			}
 		}
